add switches to force modified or standard packets
This fixes #3, if the user specifies --no-mod-packets.
Generally the point is that the autodetection of when modified packets
are present was just a wrong guess. I didn't want to try modifying it
further, as that might break files that were working before, so I've
just given up and added the switch.